requirements.txt
asgiref
certifi
crispy-bootstrap5
distlib
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
environ
filelock
Markdown
pillow
pipenv
platformdirs
psycopg2
PyJWT
setuptools
sqlparse
tzdata
virtualenv
